Refactor out WinToastTemplate into its own header file
